@@ -5,7 +5,7 @@
 
 	<groupId>io.mosip.esignet</groupId>
 	<artifactId>mock-plugin</artifactId>
-	<version>1.3.3-SNAPSHOT</version>
+	<version>1.4.0-SNAPSHOT</version>
 	<name>mock-plugin</name>
 	<packaging>jar</packaging>
 	<description>Mockup of a wrapper implementation that is used to showcase the integration with e-Signet</description>
@@ -67,44 +67,56 @@
 	</distributionManagement>
 
 	<properties>
-		<java.version>11</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-
-		<!-- maven -->
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
-		<maven.compiler.version>3.8.0</maven.compiler.version>
-		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
-		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
-		<maven.surefire.plugin.version>2.22.0</maven.surefire.plugin.version>
-		<maven.jacoco.version>0.8.5</maven.jacoco.version>
-		<maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version>
-		<maven.javadoc.version>3.2.0</maven.javadoc.version>
+		<java.version>21</java.version>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.target>21</maven.compiler.target>
+		<maven-assembly-plugin.version>3.3.0</maven-assembly-plugin.version>
+		<maven-compiler-plugin.version>2.4</maven-compiler-plugin.version>
+		<maven-surefire-plugin.version>2.22.0</maven-surefire-plugin.version>
+		<maven-gpg-plugin.version>1.5</maven-gpg-plugin.version>
+		<maven-source-plugin>2.2.1</maven-source-plugin>
+		<nexus-staging-maven-plugin>1.6.7</nexus-staging-maven-plugin>
+		<git-commit-id-plugin.version>3.0.1</git-commit-id-plugin.version>
+		<maven.jacoco.version>0.8.9</maven.jacoco.version>
+		<maven-javadoc-plugin.version>3.3.1</maven-javadoc-plugin.version>
 		<maven-shade-plugin.version>2.3</maven-shade-plugin.version>
 
-		<spring-cloud.version>Hoxton.SR8</spring-cloud.version>
-		<spring.boot.version>2.3.6.RELEASE</spring.boot.version>
+		<spring.boot.version>3.2.3</spring.boot.version>
 
-		<kernel-keymanager-service.version>1.2.2.0-SNAPSHOT</kernel-keymanager-service.version>
-		<esignet.version>1.6.1-SNAPSHOT</esignet.version>
-		<esignet-signup.version>1.2.1-SNAPSHOT</esignet-signup.version>
+		<kernel-keymanager-service.version>1.3.0-beta.2</kernel-keymanager-service.version>
+		<esignet.version>1.7.0-SNAPSHOT</esignet.version>
+		<esignet-signup.version>1.3.0-SNAPSHOT</esignet-signup.version>
 
 		<sonar.exclusions>**/dto/**,**/entity/**,**/exception/**,**/spi/**,**/advice/**,**/config/**</sonar.exclusions>
 		<sonar.cpd.exclusions>**/dto/**,**/entity/**,**/config/**</sonar.cpd.exclusions>
 	</properties>
-
 	<dependencies>
+
 		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.13.1</version>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter</artifactId>
+			<version>5.10.2</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-core</artifactId>
+			<version>5.12.0</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-junit-jupiter</artifactId>
+			<version>5.12.0</version>
+			<scope>test</scope>
+		</dependency>
+
 
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
-			<version>1.18.22</version>
+			<version>1.18.30</version>
 			<scope>compile</scope>
 		</dependency>
 
